windows: green-light bun_sys/uws_sys/threading/dns/watcher/dotenv on msvc target
Round 3 of the cross-windows leaf-crate sweep. bun_sys and bun_uws_sys now
compile clean on x86_64-pc-windows-msvc, which unblocked the next ring of
dependents; this commit clears those too. Host (linux) build stays green.

bun_sys / windows:
- mkdirat: route through open_dir_at_windows_nt_path(OnlyCreate) per
  sys.zig:829 mkdiratW (was calling a never-defined ntdll_mkdirat)
- realpath: coerce get_fd_path's &mut [u8] to &[u8]
- close_and_move_to: capture handle before self.close() consumes it
- Bun__errnoName: per-target widen for SystemErrno::init signatures
- edit_win32_binary_subsystem: drop phantom File::reader/writer; read/write
  the 4-/2-byte ints directly
- Bun__LoadLibraryBunString: match on EncodingNonAscii (what String::encoding
  actually returns), inline the latin1 zero-extend
- rename_at_w / move_opened_file_at_loose: NtCreateFileOptions /
  WindowsOpenDirOptions (the names that exist)
- spawn_in_job: assemble dwCreationFlags as DWORD (CreateProcessFlags struct
  was never ported)
- get_module_handle_from_address: cast addr to *mut c_void (extern signature)
- CreateHardLinkW debug log: use new wcslen() helper instead of WStr::from_ptr_nul
- add SRWLOCK / CONDITION_VARIABLE types + kernel32 SleepConditionVariableSRW /
  Wake{All,}ConditionVariable; re-export FILE_OPEN
- add GetProcessMemoryInfo wrapper (K32GetProcessMemoryInfo) +
  PROCESS_MEMORY_COUNTERS for bun_spawn_sys
- sys_uv: drop the as _ ambiguity on Error::new; statfs reads uv_statfs_t
  directly (no .init wrapper on Windows)
- env.rs: to_utf8_alloc returns Vec<u8>, no ?

bun_windows_sys:
- ntdll: RtlWaitOnAddress / RtlWakeAddress{Single,All} (bun_threading::Futex)
- ws2_32: addrinfo / freeaddrinfo / AF_UNIX / IPPROTO_{TCP,UDP} (bun_dns)
- NTSTATUS::TIMEOUT

bun_libuv_sys:
- uv_stat_t fields renamed to st_* to match uv.h and be field-compatible with
  libc::stat for cross-platform stat.st_mode/st_size callers (PosixStat
  conversion + lib.rs follow)

bun_errno (windows):
- S::ISREG/ISDIR/... upper-case fns taking u32, matching the POSIX S surface

bun_uws_sys:
- cfg(not(windows)) the impl PosixLoop block (every method is *mut Loop typed,
  and on Windows Loop = WindowsLoop)
- WindowsNamedPipe: link-time-dispatch shim to runtime crate (same pattern as
  UpgradedDuplex) so InternalSocket::Pipe arm dispatches
- write_fd: replace compile_error! (eager in Rust) with unreachable!()
- get_fd / fd / get_native_handle: usize SOCKET → u64 Fd backing
- from_fd: HANDLE → LIBUS_SOCKET_DESCRIPTOR cast on the windows arm

bun_threading:
- Mutex.impl_ / DebugImpl fields / WindowsImpl.srwlock pub(crate) so Condition
  can reach the SRWLOCK for SleepConditionVariableSRW
- TryAcquireSRWLockExclusive returns BOOLEAN (u8), compare against 0
- Condition: GetLastError()==TIMEOUT compares raw DWORD; current_thread_id name

bun_dns:
- platform sock:: shim re-exporting libc on posix / ws2_32 on windows; all
  addrinfo/AF_*/SOCK_*/IPPROTO_* references routed through it
- add bun_windows_sys cfg(windows) dep

bun_boringssl: local AF_INET/AF_INET6 (libc has no winsock consts)
bun_md: _strnicmp on Windows (no strncasecmp)
bun_libarchive: WStr/ZStr-agnostic deref to &[OSPathChar]
bun_dotenv: CaseInsensitiveAsciiStringArrayHashMap exists now — wire it
bun_spawn_sys: posix_spawn Action.fds is c_int (C ABI) + fd_int() helper that
  traps on Windows; pid_t = c_int locally; PeakWorkingSetSize cast
bun_watcher: resolve_path::is_parent_or_equal path; SystemErrno::init(u32) for
  the duplicate-Win32Error newtype

Remaining: bun_io (PipeReader/PipeWriter/source.rs windows arm) needs the
libuv handle-method surface (Pipe/Tty close/ref/unref/fd, ReturnCode::to_error,
TtyMode, fs_t::set_data) plus a Syscall=Tag alias — separate round.

+++ src/uws_sys/Response.rs
@@ -288,12 +288,12 @@ impl<const SSL: bool> Response<SSL> {
     pub fn get_native_handle(&mut self) -> Fd {
         #[cfg(windows)]
         {
-            // on windows uSockets exposes SOCKET
+            // on windows uSockets exposes SOCKET (uintptr-sized) as a pointer
+            // value; reinterpret as the u64 backing of `Fd` (kind=system).
             // SAFETY: uws_res_get_native_handle returns the OS SOCKET handle as a pointer.
             return Fd::from_native(unsafe {
                 c::uws_res_get_native_handle(Self::ssl_flag(), self.downcast())
-            }
-            .cast());
+            } as u64);
         }
         #[cfg(not(windows))]
         {
